+++ include/ignition/gazebo/components/JointVelocityCmd.hh
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
-*/
-#ifndef IGNITION_GAZEBO_COMPONENTS_ALTIMETER_HH_
-#define IGNITION_GAZEBO_COMPONENTS_ALTIMETER_HH_
+ */
+#ifndef IGNITION_GAZEBO_COMPONENTS_JOINTVELOCITYCMD_HH_
+#define IGNITION_GAZEBO_COMPONENTS_JOINTVELOCITYCMD_HH_
 
-#include <sdf/Sensor.hh>
+#include <vector>
 
-#include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Component.hh>
+#include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Serialization.hh>
 #include <ignition/gazebo/config.hh>
+#include <ignition/gazebo/Export.hh>
 
 namespace ignition
 {
@@ -32,10 +33,12 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief A component type that contains an altimeter sensor,
-  /// sdf::Altimeter, information.
-  using Altimeter = Component<sdf::Sensor, class AltimeterTag>;
-  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.Altimeter", Altimeter)
+  /// \brief Base class which can be extended to add serialization
+  using JointVelocityCmd = Component<std::vector<double>,
+        class JointVelocityCmdTag>;
+
+  IGN_GAZEBO_REGISTER_COMPONENT(
+      "ign_gazebo_components.JointVelocityCmd", JointVelocityCmd)
 }
 }
 }